@@ -1,8 +1,8 @@
 {
     "name": "@solana/wallet-adapter-create-react-app-starter",
-    "version": "0.1.15",
+    "version": "0.1.30",
     "author": "Solana Maintainers <maintainers@solana.foundation>",
-    "repository": "https://github.com/solana-labs/wallet-adapter",
+    "repository": "https://github.com/anza-xyz/wallet-adapter",
     "license": "Apache-2.0",
     "publishConfig": {
         "access": "public"
@@ -28,10 +28,10 @@
     },
     "dependencies": {
         "@solana/wallet-adapter-base": "workspace:^",
-        "@solana/wallet-adapter-react": "workspace:^0.15.30",
-        "@solana/wallet-adapter-react-ui": "workspace:^0.9.29",
-        "@solana/wallet-adapter-wallets": "workspace:^0.19.15",
-        "@solana/web3.js": "^1.73.2",
+        "@solana/wallet-adapter-react": "workspace:^",
+        "@solana/wallet-adapter-react-ui": "workspace:^",
+        "@solana/wallet-adapter-wallets": "workspace:^",
+        "@solana/web3.js": "^1.77.3",
         "react": "^18.2.0",
         "react-app-rewired": "^2.2.1",
         "react-dom": "^18.2.0",
@@ -43,14 +43,20 @@
         "@testing-library/react": "^13.4.0",
         "@testing-library/user-event": "^14.4.3",
         "@types/jest": "^28.1.8",
-        "@types/react": "^18.0.28",
-        "@types/react-dom": "^18.0.11",
-        "@types/testing-library__jest-dom": "^5.14.5",
+        "@types/react": "^18.2.13",
+        "@types/react-dom": "^18.2.6",
+        "@types/testing-library__jest-dom": "^5.14.6",
+        "browserify-zlib": "^0.2.0",
+        "crypto-browserify": "^3.12.0",
+        "https-browserify": "^1.0.0",
         "process": "^0.11.10",
         "shx": "^0.3.4",
         "source-map-loader": "^4.0.1",
+        "stream-browserify": "^3.0.0",
+        "stream-http": "^3.2.0",
         "typescript": "~4.7.4",
-        "webpack": "^5.75.0"
+        "url": "^0.11.1",
+        "webpack": "^5.88.0"
     },
     "eslintConfig": {
         "extends": [
