@@ -1,10 +1,10 @@
 import React, { StrictMode } from 'react';
-import ReactDOM from 'react-dom';
+import { createRoot } from 'react-dom/client';
 import { App } from './App';
 
-ReactDOM.render(
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+createRoot(document.getElementById('app')!).render(
     <StrictMode>
         <App />
-    </StrictMode>,
-    document.getElementById('app')
+    </StrictMode>
 );
