@@ -1,9 +1,9 @@
 import { Button } from '@mui/material';
+import { ed25519 } from '@noble/curves/ed25519';
 import { useWallet } from '@solana/wallet-adapter-react';
 import bs58 from 'bs58';
 import type { FC } from 'react';
 import React, { useCallback } from 'react';
-import { sign } from 'tweetnacl';
 import { useNotify } from './notify';
 
 export const SignMessage: FC = () => {
@@ -15,14 +15,17 @@ export const SignMessage: FC = () => {
             if (!publicKey) throw new Error('Wallet not connected!');
             if (!signMessage) throw new Error('Wallet does not support message signing!');
 
-            const message = new TextEncoder().encode('Hello, world!');
+            const message = new TextEncoder().encode(
+                `${
+                    window.location.host
+                } wants you to sign in with your Solana account:\n${publicKey.toBase58()}\n\nPlease sign in.`
+            );
             const signature = await signMessage(message);
-            if (!sign.detached.verify(message, signature, publicKey.toBytes()))
-                throw new Error('Message signature invalid!');
 
+            if (!ed25519.verify(signature, message, publicKey.toBytes())) throw new Error('Message signature invalid!');
             notify('success', `Message signature: ${bs58.encode(signature)}`);
         } catch (error: any) {
-            notify('error', `Message signing failing: ${error?.message}`);
+            notify('error', `Sign Message failed: ${error?.message}`);
         }
     }, [publicKey, signMessage, notify]);
 
