@@ -1,52 +1,14 @@
-import { useWallet } from '@solana/wallet-adapter-react';
 import type { ButtonProps } from 'antd';
-import { Button } from 'antd';
-import type { FC, MouseEventHandler } from 'react';
-import React, { useCallback, useMemo } from 'react';
-import { WalletIcon } from './WalletIcon.js';
+import React from 'react';
+import { BaseWalletConnectButton } from './BaseWalletConnectButton.js';
 
-export const WalletConnectButton: FC<ButtonProps> = ({
-    type = 'primary',
-    size = 'large',
-    htmlType = 'button',
-    children,
-    disabled,
-    onClick,
-    ...props
-}) => {
-    const { wallet, connect, connecting, connected } = useWallet();
+const LABELS = {
+    connecting: 'Connecting ...',
+    connected: 'Connected',
+    'has-wallet': 'Connect',
+    'no-wallet': 'Connect Wallet',
+} as const;
 
-    const handleClick: MouseEventHandler<HTMLButtonElement> = useCallback(
-        (event) => {
-            if (onClick) onClick(event);
-            // eslint-disable-next-line @typescript-eslint/no-empty-function
-            if (!event.defaultPrevented)
-                connect().catch(() => {
-                    // Silently catch because any errors are caught by the context `onError` handler
-                });
-        },
-        [onClick, connect]
-    );
-
-    const content = useMemo(() => {
-        if (children) return children;
-        if (connecting) return 'Connecting ...';
-        if (connected) return 'Connected';
-        if (wallet) return 'Connect';
-        return 'Connect Wallet';
-    }, [children, connecting, connected, wallet]);
-
-    return (
-        <Button
-            onClick={handleClick}
-            disabled={disabled || !wallet || connecting || connected}
-            icon={<WalletIcon wallet={wallet} />}
-            type={type}
-            size={size}
-            htmlType={htmlType}
-            {...props}
-        >
-            {content}
-        </Button>
-    );
-};
+export function WalletConnectButton(props: ButtonProps) {
+    return <BaseWalletConnectButton {...props} labels={LABELS} />;
+}
