@@ -3,7 +3,7 @@ import type { DetailedHTMLProps, FC, ImgHTMLAttributes } from 'react';
 import React from 'react';
 
 export interface WalletIconProps extends DetailedHTMLProps<ImgHTMLAttributes<HTMLImageElement>, HTMLImageElement> {
-    wallet: Wallet | null;
+    wallet: { adapter: Pick<Wallet['adapter'], 'icon' | 'name'> } | null;
 }
 
 export const WalletIcon: FC<WalletIconProps> = ({ wallet, ...props }) => {
