@@ -1,17 +1,10 @@
 import type { ButtonProps } from 'antd';
-import { Button } from 'antd';
 import type { FC, MouseEventHandler } from 'react';
 import React, { useCallback } from 'react';
+import { BaseWalletConnectionButton } from './BaseWalletConnectionButton.js';
 import { useWalletModal } from './useWalletModal.js';
 
-export const WalletModalButton: FC<ButtonProps> = ({
-    children = 'Select Wallet',
-    type = 'primary',
-    size = 'large',
-    htmlType = 'button',
-    onClick,
-    ...props
-}) => {
+export const WalletModalButton: FC<ButtonProps> = ({ children = 'Select Wallet', onClick, ...props }) => {
     const { setVisible } = useWalletModal();
 
     const handleClick: MouseEventHandler<HTMLButtonElement> = useCallback(
@@ -23,8 +16,8 @@ export const WalletModalButton: FC<ButtonProps> = ({
     );
 
     return (
-        <Button onClick={handleClick} type={type} size={size} htmlType={htmlType} {...props}>
+        <BaseWalletConnectionButton {...props} onClick={handleClick}>
             {children}
-        </Button>
+        </BaseWalletConnectionButton>
     );
 };
