@@ -1,39 +1,14 @@
-import { useWallet } from '@solana/wallet-adapter-react';
-import type { FC, MouseEventHandler } from 'react';
-import React, { useCallback, useMemo } from 'react';
+import React from 'react';
+import { BaseWalletConnectButton } from './BaseWalletConnectButton.js';
 import type { ButtonProps } from './Button.js';
-import { Button } from './Button.js';
-import { WalletIcon } from './WalletIcon.js';
 
-export const WalletConnectButton: FC<ButtonProps> = ({ children, disabled, onClick, ...props }) => {
-    const { wallet, connect, connecting, connected } = useWallet();
+const LABELS = {
+    connecting: 'Connecting ...',
+    connected: 'Connected',
+    'has-wallet': 'Connect',
+    'no-wallet': 'Connect Wallet',
+} as const;
 
-    const handleClick: MouseEventHandler<HTMLButtonElement> = useCallback(
-        (event) => {
-            if (onClick) onClick(event);
-            // eslint-disable-next-line @typescript-eslint/no-empty-function
-            if (!event.defaultPrevented) connect().catch(() => {});
-        },
-        [onClick, connect]
-    );
-
-    const content = useMemo(() => {
-        if (children) return children;
-        if (connecting) return 'Connecting ...';
-        if (connected) return 'Connected';
-        if (wallet) return 'Connect';
-        return 'Connect Wallet';
-    }, [children, connecting, connected, wallet]);
-
-    return (
-        <Button
-            className="wallet-adapter-button-trigger"
-            disabled={disabled || !wallet || connecting || connected}
-            startIcon={wallet ? <WalletIcon wallet={wallet} /> : undefined}
-            onClick={handleClick}
-            {...props}
-        >
-            {content}
-        </Button>
-    );
-};
+export function WalletConnectButton(props: ButtonProps) {
+    return <BaseWalletConnectButton {...props} labels={LABELS} />;
+}
