@@ -1,7 +1,7 @@
 import type { FC, MouseEvent } from 'react';
 import React, { useCallback } from 'react';
 import type { ButtonProps } from './Button.js';
-import { Button } from './Button.js';
+import { Button as BaseWalletConnectionButton } from './Button.js';
 import { useWalletModal } from './useWalletModal.js';
 
 export const WalletModalButton: FC<ButtonProps> = ({ children = 'Select Wallet', onClick, ...props }) => {
@@ -16,8 +16,8 @@ export const WalletModalButton: FC<ButtonProps> = ({ children = 'Select Wallet',
     );
 
     return (
-        <Button className="wallet-adapter-button-trigger" onClick={handleClick} {...props}>
+        <BaseWalletConnectionButton {...props} className="wallet-adapter-button-trigger" onClick={handleClick}>
             {children}
-        </Button>
+        </BaseWalletConnectionButton>
     );
 };
