@@ -1,16 +1,11 @@
 export * from '@solana/wallet-adapter-alpha';
 export * from '@solana/wallet-adapter-avana';
-export * from '@solana/wallet-adapter-backpack';
 export * from '@solana/wallet-adapter-bitkeep';
 export * from '@solana/wallet-adapter-bitpie';
-export * from '@solana/wallet-adapter-blocto';
-export * from '@solana/wallet-adapter-brave';
-export * from '@solana/wallet-adapter-censo';
 export * from '@solana/wallet-adapter-clover';
 export * from '@solana/wallet-adapter-coin98';
 export * from '@solana/wallet-adapter-coinbase';
 export * from '@solana/wallet-adapter-coinhub';
-export * from '@solana/wallet-adapter-exodus';
 export * from '@solana/wallet-adapter-fractal';
 export * from '@solana/wallet-adapter-foxwallet';
 export * from '@solana/wallet-adapter-glow';
@@ -19,7 +14,6 @@ export * from '@solana/wallet-adapter-hyperpay';
 export * from '@solana/wallet-adapter-keystone';
 export * from '@solana/wallet-adapter-krystal';
 export * from '@solana/wallet-adapter-ledger';
-// export * from '@solana/wallet-adapter-magiceden';
 export * from '@solana/wallet-adapter-mathwallet';
 export * from '@solana/wallet-adapter-neko';
 export * from '@solana/wallet-adapter-nightly';
@@ -31,15 +25,13 @@ export * from '@solana/wallet-adapter-safepal';
 export * from '@solana/wallet-adapter-saifu';
 export * from '@solana/wallet-adapter-salmon';
 export * from '@solana/wallet-adapter-sky';
-export * from '@solana/wallet-adapter-slope';
 export * from '@solana/wallet-adapter-solflare';
-export * from '@solana/wallet-adapter-sollet';
 export * from '@solana/wallet-adapter-solong';
 export * from '@solana/wallet-adapter-spot';
-export * from '@solana/wallet-adapter-strike';
 export * from '@solana/wallet-adapter-tokenary';
 export * from '@solana/wallet-adapter-tokenpocket';
 export * from '@solana/wallet-adapter-torus';
+export * from '@solana/wallet-adapter-trezor';
 export * from '@solana/wallet-adapter-trust';
 export * from '@solana/wallet-adapter-unsafe-burner';
 export * from '@solana/wallet-adapter-walletconnect';
